chore(deps): update dependencies in bun.lock and package.json
- Added @orpc/contract and viem dependencies, both at version 1.7.4 and 2.32.1 respectively.
- Updated prometheus-node-exporter Helm chart version from 4.47.1 to 4.47.2.
- Changed linker setting in bunfig.toml to "isolated".
- Removed deprecated tanstack-form component files.

These updates enhance the project's dependency management and ensure compatibility with the latest versions.
