+++ bunfig.toml
@@ -1,6 +1,6 @@
 [install]
 exact = true
-#linker = "isolated"
+linker = "isolated"
 
 [console]
 depth = 5
