+++ kit/charts/atk/charts/observability/README.md
@@ -20,7 +20,7 @@ A Helm chart for the observability components
 | https://grafana.github.io/helm-charts | tempo | 1.23.2 |
 | https://kubernetes-sigs.github.io/metrics-server/ | metrics-server | 3.12.2 |
 | https://prometheus-community.github.io/helm-charts | kube-state-metrics | 6.1.0 |
-| https://prometheus-community.github.io/helm-charts | prometheus-node-exporter | 4.47.1 |
+| https://prometheus-community.github.io/helm-charts | prometheus-node-exporter | 4.47.2 |
 | https://victoriametrics.github.io/helm-charts/ | victoria-metrics-single | 0.23.0 |
 
 ## Values
