+++ kit/dapp/package.json
@@ -69,6 +69,7 @@
     "@fontsource-variable/figtree": "5.2.8",
     "@fontsource-variable/roboto-mono": "5.2.6",
     "@orpc/client": "1.7.4",
+    "@orpc/contract": "1.7.4",
     "@orpc/json-schema": "1.7.4",
     "@orpc/openapi": "1.7.4",
     "@orpc/react-query": "1.7.4",
@@ -159,6 +160,7 @@
     "tw-animate-css": "1.3.5",
     "usehooks-ts": "3.1.1",
     "vaul": "1.1.2",
+    "viem": "2.32.1",
     "zod": "4.0.5"
   }
 }
